LPS-104152 Remove old DDM JSPs and MVC actions
